Refactor so tabs are at the page level for loading state
